added three feature test files for diversified user stories, 204 assertions passing
