+++ test/features/user_can_view_event_specific_details_test.rb
@@ -39,15 +39,4 @@ class UserCanViewEventSpecificDetails < FeatureTest
     end
   end
 
-  def test_user_cannot_view_details_of_an_undefined_event
-    register_turing_and_send_multiple_payloads
-
-    visit '/sources/turing/events/socialZZYZX'
-
-    within 'h1' do
-      assert page.has_content?('event: socialZZYZX has not been defined for this application')
-    end
-
-    assert page.has_css?("a[href~='/sources/turing/events']")
-  end
 end
